bento: start extracting AMP specific bits
